+++ cli/src/commands/squash.rs
@@ -25,6 +25,7 @@ use jj_lib::rewrite;
 use jj_lib::rewrite::CommitWithSelection;
 use jj_lib::rewrite::merge_commit_trees;
 use jj_lib::rewrite::rebase_commit;
+use jj_lib::trailer::Trailer;
 use pollster::FutureExt as _;
 use tracing::instrument;
 
@@ -37,7 +38,8 @@ use crate::command_error::CommandError;
 use crate::command_error::user_error;
 use crate::command_error::user_error_with_hint;
 use crate::complete;
-use crate::description_util::add_trailers;
+use crate::description_util::add_config_trailers;
+use crate::description_util::add_config_trailers_with_extras;
 use crate::description_util::combine_messages_for_editing;
 use crate::description_util::description_template;
 use crate::description_util::edit_description;
@@ -160,6 +162,16 @@ pub(crate) struct SquashArgs {
     )]
     tool: Option<String>,
 
+    /// Set the authors of the squashed commits as co-authors of the
+    /// destination commit
+    #[arg(long, value_enum, group = "co-authors")]
+    co_authors: bool,
+
+    /// Don't set the authors of the squashed commits as co-authors of the
+    /// destination commit
+    #[arg(long, value_enum, group = "co-authors")]
+    no_co_authors: bool,
+
     /// Move only changes to these paths (instead of all paths)
     #[arg(
         value_name = "FILESETS",
@@ -273,7 +285,7 @@ pub(crate) fn cmd_squash(
         let mut description = join_message_paragraphs(&args.message_paragraphs);
         if !description.is_empty() {
             commit_builder.set_description(description);
-            description = add_trailers(ui, &tx, &commit_builder)?;
+            description = add_config_trailers(ui, &tx, &commit_builder)?;
         }
         commit_builder.set_description(&description);
         let commit = commit_builder.write(tx.repo_mut())?;
@@ -323,12 +335,35 @@ pub(crate) fn cmd_squash(
     let description = SquashedDescription::from_args(args);
 
     let source_commits = select_diff(&tx, &sources, &destination, &matcher, &diff_selector)?;
+    let use_co_authors_trailers = if args.co_authors {
+        true
+    } else if args.no_co_authors {
+        false
+    } else {
+        tx.settings().get_bool("squash.co-authors")?
+    };
     if let Some(squashed) = rewrite::squash_commits(
         tx.repo_mut(),
         &source_commits,
         &destination,
         args.keep_emptied,
     )? {
+        let co_author_trailers: Vec<_> = if use_co_authors_trailers {
+            squashed
+                .abandoned_commits
+                .iter()
+                .filter(|commit| {
+                    commit.author().name != destination.author().name
+                        || commit.author().email != destination.author().email
+                })
+                .map(|commit| Trailer {
+                    key: "Co-authored-by".to_string(),
+                    value: format!("{} <{}>", commit.author().name, commit.author().email),
+                })
+                .collect()
+        } else {
+            vec![]
+        };
         let mut commit_builder = squashed.commit_builder.detach();
         let new_description = match description {
             SquashedDescription::Exact(description) => {
@@ -336,7 +371,7 @@ pub(crate) fn cmd_squash(
                     description
                 } else {
                     commit_builder.set_description(description);
-                    add_trailers(ui, &tx, &commit_builder)?
+                    add_config_trailers_with_extras(ui, &tx, &commit_builder, &co_author_trailers)?
                 }
             }
             SquashedDescription::UseDestination => {
@@ -344,7 +379,7 @@ pub(crate) fn cmd_squash(
                     destination.description().to_owned()
                 } else {
                     commit_builder.set_description(destination.description());
-                    add_trailers(ui, &tx, &commit_builder)?
+                    add_config_trailers_with_extras(ui, &tx, &commit_builder, &co_author_trailers)?
                 }
             }
             SquashedDescription::Combine => {
@@ -357,7 +392,12 @@ pub(crate) fn cmd_squash(
                         description
                     } else {
                         commit_builder.set_description(description);
-                        add_trailers(ui, &tx, &commit_builder)?
+                        add_config_trailers_with_extras(
+                            ui,
+                            &tx,
+                            &commit_builder,
+                            &co_author_trailers,
+                        )?
                     }
                 } else {
                     let combined = combine_messages_for_editing(
@@ -370,9 +410,10 @@ pub(crate) fn cmd_squash(
                             Some(&destination)
                         },
                         &commit_builder,
+                        &co_author_trailers,
                     )?;
                     // It's weird that commit.description() contains "JJ: " lines, but works.
-                    commit_builder.set_description(combined);
+                    commit_builder.set_description(&combined);
                     let temp_commit = commit_builder.write_hidden()?;
                     let intro = "Enter a description for the combined commit.";
                     let template = description_template(ui, &tx, intro, &temp_commit)?;